🎨 [Structure]: #472 task-links ドメインの index.ts を内部モジュールへ分割
644 行に肥大した index.ts を責務単位の 5 モジュールへ分割し、
index.ts は公開 API（型 re-export + TaskLinks companion の組み立て）のみにする。

- types.ts            TaskLinks 型
- linkOperation.ts    LinkOperation 型 + 適用（applyLinkOperationsToTask / grouping）
- plan.ts             planAddLink / planRemoveLink + inverse rollback 計算
- candidates.ts       buildAddLinkCandidates / buildCreateLinkCandidates
- removeLinkedTask.ts task-deleted 時のリンク掃除

ロジックは無変更（移動のみ）。公開 API・import パス（@/domains/task-links）も不変で、
全テスト 2529 件が無改修 PASS。

Refs #472
